+++ programs/fixed-term/src/tickets/instructions/stake_tickets.rs
@@ -86,7 +86,7 @@ pub fn handler(ctx: Context<StakeTickets>, params: StakeTicketsParams) -> Result
         amount,
     )?;
 
-    let matures_at = Clock::get()?.unix_timestamp + ctx.accounts.market.load()?.lend_tenor;
+    let matures_at = Clock::get()?.unix_timestamp + ctx.accounts.market.load()?.lend_tenor as i64;
 
     // Mint a deposit for their burned tokens
     *ctx.accounts.deposit = TermDeposit {
